add ref collection endpoint /api/guildadd
redirects to /thanks (to be a "thank you for adding the bot, would you like to try out the *dashboard* or perhaps *view commands*?" thing) and collects ref (either a dbots website, github, discord, or the website itself). referrals are collected via a Referral object, one for each type of referral, stored in an object. Who knows if I'll ever have a practical use for this data. I guess I'll show it on the dev dashboard or whatever. Lol.